Reconfigure connectionUrl
Due to the way getConnectionUrl() works, i.e. it will look for a
developerConnection is the connection is empty, and the fact that
-DconnectionType can specify which SCM url should be used, both
connectionUrl and developerConnectionUrl need reconfiguring in the mojo.

Also added debug calls so that mvn -X displays what the reconfiguration
looks like. If you run this in a directory with a pom.xml already the
scm plugin will be configured with the current directores scm section
which is confusing in the debug output.