Fix snapshot handling in ITs
The default updatePolicy for snapshots is daily, which means that local
changes get checked once a day. This needs to be set to always so that
whenever a local build is done it gets put into the IT repo.

apache.snapshots should not be used.
Otherwise snapshots get pulled down and overwrite your local changes.